fix(gateway): keep the trash root policy off the plugin SDK surface
The plugin SDK infra barrel re-exports the fs-safe facade wholesale, so the
new helper grew the budgeted public surface. Own it with the cleanup code in
src/commands/cleanup-utils.ts instead; both delete transports import it there.
